email interactions

+++ src/org/thoughtcrime/securesms/ConversationListFragment.java
@@ -244,7 +244,12 @@ public class ConversationListFragment extends Fragment
       @Override
       protected void executeAction(@Nullable Void parameter) {
         for (long threadId : selectedConversations) {
-          dcContext.archiveChat((int)threadId, !archive? 1 : 0);
+          if (threadId==DcChat.DC_CHAT_ID_DEADDROP) {
+            dcContext.marknoticedContact(getListAdapter().getDeaddropContactId());
+          }
+          else {
+            dcContext.archiveChat((int)threadId, !archive? 1 : 0);
+          }
         }
       }
 
@@ -285,7 +290,12 @@ public class ConversationListFragment extends Fragment
           @Override
           protected Void doInBackground(Void... params) {
             for (long threadId : selectedConversations) {
-              dcContext.deleteChat((int)threadId);
+              if (threadId==DcChat.DC_CHAT_ID_DEADDROP) {
+                dcContext.marknoticedContact(getListAdapter().getDeaddropContactId());
+              }
+              else {
+                dcContext.deleteChat((int) threadId);
+              }
             }
             return null;
           }
@@ -327,6 +337,8 @@ public class ConversationListFragment extends Fragment
     return new DcChatlistLoader(getActivity(), listflags, queryFilter.isEmpty()? null : queryFilter, 0);
   }
 
+
+  boolean forceListRedraw;
   @Override
   public void onLoadFinished(Loader<DcChatlist> arg0, DcChatlist chatlist) {
     if (chatlist.getCnt() <= 0 && TextUtils.isEmpty(queryFilter) && !archive) {
@@ -346,7 +358,18 @@ public class ConversationListFragment extends Fragment
       fab.stopPulse();
     }
 
+    // this hack is needed as otherwise, for whatever reason,
+    // swiped contact request show an empty item if there pops up a new contact request imediately.
+    // anyone who wants to invesigate to this is very welcome :)
+    if (forceListRedraw) {
+      list.setLayoutManager(null);
+      list.getRecycledViewPool().clear();
+      list.setLayoutManager(new LinearLayoutManager(getActivity()));
+      forceListRedraw = false;
+    }
+
     getListAdapter().changeData(chatlist);
+
   }
 
   @Override
@@ -515,6 +538,7 @@ public class ConversationListFragment extends Fragment
         if (threadId==DcChat.DC_CHAT_ID_DEADDROP) {
           int contactId = ((ConversationListItem)viewHolder.itemView).getContactId();
           dcContext.marknoticedContact(contactId);
+          forceListRedraw = true;
           return;
         }
